新增了获取当前计时的函数，使用了time.h中的clock_gettime函数，具有很好的精度和可移植性;将rendering.c中的timerun_print函数接收的毫秒数改为long类型

+++ rendering.h
@@ -8,7 +8,7 @@ void init_TUI();
 void init_gameboard (Table* table, WINDOW** gmbrd_win, WINDOW** plygrnd_win, WINDOW** timrun_win, WINDOW** msg_win, WINDOW** tips_win);
 void maze_render (Table* table, WINDOW* playground_win);
 void coordinate_screen_move (WINDOW* win, Coordinate* coordinate);
-void timerun_print(WINDOW* win, WINDOW* tim_win, int time_ms);
+void timerun_print (WINDOW* win, WINDOW* tim_win, long time_ms);
 void message_tips_print (WINDOW* win, char* str);
 void cursor_reset ();
